blit: the two GPU fast paths are a rail's, and the rail is asked through Backend
`execute_blit` is a host copy on both rails, and every function it calls is
backend-neutral. What was not neutral were two fast paths — a whole-plane
type-11 copy out of a resident, and a whole-plane type-11 to guest-linear one —
plus a resident census. They lived in the shared module: one behind two
same-named definitions a `cfg` chose between (one of them a `{ None }` stub),
one behind an inline `cfg` block at its call site, and the census behind a
`cfg`-else-`let _ = args`.

* `blit_exec::vulkan` is the rail. The two fast paths, the `T2tGvaRefusal`
  vocabulary, the GVA plane resolve and the resident census move into it, along
  with the one test that reaches `gpu_t2t_gva_plane`.
* `Backend` gains the three, each with a default that is the honest Metal
  answer: decline the optimisation, record no reading. The stub's own doc had
  already argued for that shape — "spelled as a fall-through rather than as a
  `cfg` at the call site so the whole-surface form reads the same on both
  backends" — and the trait is where that shape belongs.
* The shared admissibility logic — `GpuPlane`, `GpuPlaneRefusal`,
  `gpu_whole_plane_admissible`, `gpu_whole_plane_destination` — deliberately
  stays in the shared module. It is pure arithmetic over decoded guest facts,
  its tests run on every arm, and moving it into a rail would put those tests on
  one.
* `Backend` becomes `pub(crate)`. Its vocabulary is the device's own resolved
  state — a `DrawEncodeRequest`, a resolved blit endpoint — and nothing outside
  this crate implements or calls a backend, so a public trait would only mean
  publishing those types to keep a signature legal.

No behavior change on either arm. Metal 1315 pass, Vulkan 2173 pass, both
serial; feature-matrix green with per-arm counts held.
